+++ sample-apps/react-native/dogfood/App.tsx
@@ -14,29 +14,26 @@ import {
 } from './src/contexts/AppContext';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import {
-  setFirebaseHandler,
-  setForegroundService,
-} from './src/modules/push/android';
-import { useIosPushEffect } from './src/hooks/useIosPushEffect';
-import { Platform } from 'react-native';
-import { useCallKeepEffect } from './src/hooks/useCallkeepEffect';
-import { navigationRef } from './src/utils/staticNavigationUtils';
+  StaticNavigationService,
+  navigationRef,
+} from './src/utils/staticNavigationUtils';
 import Logger from 'react-native-webrtc/src/Logger';
 import { Meeting } from './src/navigators/Meeting';
 import { Call } from './src/navigators/Call';
 import { VideoWrapper } from './src/components/VideoWrapper';
 import LoginScreen from './src/screens/LoginScreen';
-import { ChooseAppModeScreen } from './src/screens/ChooseAppModeScreen';
+import { ChatWrapper } from './src/components/ChatWrapper';
+import { AppMode } from './src/navigators/AppMode';
+import { setPushConfig } from './src/utils/setPushConfig';
 
 // @ts-expect-error
 Logger.enable(false);
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
-if (Platform.OS === 'android') {
-  setFirebaseHandler();
-  setForegroundService();
-}
+// All the push notification related config must be done as soon the app starts
+// since the app can be opened from a dead state through a push notification
+setPushConfig();
 
 const StackNavigator = () => {
   const appMode = useAppGlobalStoreValue((store) => store.appMode);
@@ -45,8 +42,6 @@ const StackNavigator = () => {
   const setState = useAppGlobalStoreSetState();
 
   useProntoLinkEffect();
-  useIosPushEffect();
-  useCallKeepEffect();
 
   let mode;
   switch (appMode) {
@@ -72,7 +67,7 @@ const StackNavigator = () => {
       mode = (
         <Stack.Screen
           name="ChooseAppMode"
-          component={ChooseAppModeScreen}
+          component={AppMode}
           options={{ headerShown: false }}
         />
       );
@@ -80,6 +75,9 @@ const StackNavigator = () => {
   }
 
   useEffect(() => {
+    if (!(username && userImageUrl)) {
+      return;
+    }
     const subscription = prontoCallId$.subscribe((prontoCallId) => {
       if (prontoCallId) {
         setState({ appMode: 'Meeting' });
@@ -87,7 +85,16 @@ const StackNavigator = () => {
     });
 
     return () => subscription.unsubscribe();
-  }, [setState]);
+  }, [setState, username, userImageUrl]);
+
+  useEffect(() => {
+    if (username && userImageUrl) {
+      StaticNavigationService.authenticationInfo = {
+        username,
+        userImageUrl,
+      };
+    }
+  }, [username, userImageUrl]);
 
   if (!(username && userImageUrl)) {
     return <LoginScreen />;
@@ -95,7 +102,9 @@ const StackNavigator = () => {
 
   return (
     <VideoWrapper>
-      <Stack.Navigator>{mode}</Stack.Navigator>
+      <ChatWrapper>
+        <Stack.Navigator>{mode}</Stack.Navigator>
+      </ChatWrapper>
     </VideoWrapper>
   );
 };