+++ sample-apps/react-native/dogfood/src/components/MeetingUI.tsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import {
-  ActiveCall,
-  ActiveCallProps,
-  useCall,
-} from '@stream-io/video-react-native-sdk';
+import { ActiveCall, useCall } from '@stream-io/video-react-native-sdk';
 import { MeetingStackParamList, ScreenTypes } from '../../types';
 import { Button, SafeAreaView, StyleSheet, Text, View } from 'react-native';
-import { theme } from '@stream-io/video-react-native-sdk/dist/src/theme';
-import { ParticipantListButtons } from '../components/ParticipantListButtons';
 import { LobbyViewComponent } from './LobbyViewComponent';
+import { useUnreadCount } from '../hooks/useUnreadCount';
+import { useChannelWatch } from '../hooks/useChannelWatch';
+import { appTheme } from '../theme';
 
 type Props = NativeStackScreenProps<
   MeetingStackParamList,
@@ -19,7 +16,6 @@ type Props = NativeStackScreenProps<
   show: ScreenTypes;
   setShow: React.Dispatch<React.SetStateAction<ScreenTypes>>;
 };
-type Mode = NonNullable<ActiveCallProps['mode']>;
 
 export const MeetingUI = ({
   callId,
@@ -28,8 +24,9 @@ export const MeetingUI = ({
   show,
   setShow,
 }: Props) => {
-  const [selectedMode, setSelectedMode] = React.useState<Mode>('grid');
   const call = useCall();
+  const channelWatched = useChannelWatch();
+  const unreadBadgeCountIndicator = useUnreadCount({ channelWatched });
 
   const returnToHomeHandler = () => {
     navigation.navigate('JoinMeetingScreen');
@@ -60,11 +57,14 @@ export const MeetingUI = ({
   } else {
     return (
       <SafeAreaView style={styles.wrapper}>
-        <ParticipantListButtons
-          selectedMode={selectedMode}
-          setMode={setSelectedMode}
+        <ActiveCall
+          chatButton={{
+            onPressHandler: () => {
+              navigation.navigate('ChatScreen', { callId: callId });
+            },
+            unreadBadgeCountIndicator,
+          }}
         />
-        <ActiveCall mode={selectedMode} />
       </SafeAreaView>
     );
   }
@@ -74,21 +74,21 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     justifyContent: 'center',
-    backgroundColor: theme.light.static_grey,
+    backgroundColor: appTheme.colors.static_grey,
   },
   wrapper: {
     flex: 1,
-    backgroundColor: theme.light.static_grey,
+    backgroundColor: appTheme.colors.static_grey,
   },
   buttonsContainer: {
     flexDirection: 'row',
     justifyContent: 'flex-end',
-    paddingVertical: 4,
-    paddingHorizontal: 16,
+    paddingVertical: appTheme.spacing.xs,
+    paddingHorizontal: appTheme.spacing.lg,
   },
   errorText: {
     fontSize: 30,
-    color: 'red',
+    color: appTheme.colors.error,
     textAlign: 'center',
   },
 });