+++ sample-apps/react-native/dogfood/src/utils/staticNavigationUtils.ts
@@ -1,35 +1,42 @@
-import {
-  StackActions,
-  createNavigationContainerRef,
-} from '@react-navigation/native';
+import { createNavigationContainerRef } from '@react-navigation/native';
 
 import { RootStackParamList } from '../../types';
 
 export const navigationRef = createNavigationContainerRef<RootStackParamList>();
 
-export const StaticNavigationService = {
-  navigate(name: string, params: Object | undefined = undefined) {
+export class StaticNavigationService {
+  static authenticationInfo:
+    | { username: string; userImageUrl: string }
+    | undefined = undefined;
+
+  /**
+   * Navigate to a route in the Root StackNavigator
+   */
+  static navigate<RouteName extends keyof RootStackParamList>(
+    name: RouteName,
+    params: RootStackParamList[RouteName] | undefined = undefined,
+  ) {
     if (navigationRef.isReady()) {
-      const currentRoute = navigationRef.getCurrentRoute();
-      if (currentRoute?.name === name) {
-        navigationRef.dispatch(StackActions.replace(name, params));
-      } else {
-        // @ts-ignore
-        navigationRef.navigate(name, params);
-      }
+      // @ts-ignore
+      navigationRef.navigate(name, params);
     }
-  },
-  goBack() {
+  }
+  static goBack() {
     if (navigationRef.isReady()) {
       navigationRef.goBack();
     }
-  },
-};
+  }
+}
 
+/**
+ * Run the navigation logic with StaticNavigationService
+ * This is used to run the navigation logic from root level even before the navigation is ready
+ * @param callback The navigation callback that calls the methods of StaticNavigationService
+ */
 export const RunStaticNavigation = (callback: () => void) => {
   const intervalId = setInterval(async () => {
-    // TODO: check if user is authenticated
-    if (navigationRef.isReady()) {
+    // run only when the navigation is ready and the user is authenticated
+    if (navigationRef.isReady() && StaticNavigationService.authenticationInfo) {
       clearInterval(intervalId);
       callback();
     }